refactor: extract SC search hook and utilities from MetaEditorPage
- Extract useSoundCloudSearch hook: wraps handleScSearch in useCallback,
  fixing the stale closure bug where handleScSearch was called from a
  useEffect without being in useCallback (issue #45 primary concern).
  Moves 5 state variables and debounced search logic out of the component.

- Extract utility functions (parseComment, serializeComment,
  stripQueryParams, scReleaseDate, formatFileSize) to utils.ts,
  making them testable and reusable.

- Page reduced from 1492 to 1374 lines.

Closes #45
